requirements.txt
envparse==0.2.0
fastapi==0.104.1
python-telegram-bot==20.7
uvicorn==0.24.0
varcache==0.1.0
